+++ source/ProcessManager.Orchestrations.Abstractions/Processes/BRS_021/ForwardMeteredData/V1/Model/ForwardMeteredDataAcceptedV1.cs
@@ -20,17 +20,23 @@ namespace Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes
 /// <summary>
 /// A model containing the data required for notifying market actors when new metered data has been accepted.
 /// </summary>
-public record MeteredDataForMeteringPointAcceptedV1(
-    string MessageId,
+public record ForwardMeteredDataAcceptedV1(
+    string OriginalActorMessageId,
+    string OriginalTransactionId,
     string MeteringPointId,
     MeteringPointType MeteringPointType,
-    string OriginalTransactionId,
     string ProductNumber,
     MeasurementUnit MeasureUnit,
     DateTimeOffset RegistrationDateTime,
     Resolution Resolution,
     DateTimeOffset StartDateTime,
     DateTimeOffset EndDateTime,
-    IReadOnlyCollection<AcceptedEnergyObservation> AcceptedEnergyObservations,
-    IReadOnlyCollection<MarketActorRecipient> MarketActorRecipients)
-    : IInputParameterDto;
+    IReadOnlyCollection<ForwardMeteredDataAcceptedV1.AcceptedEnergyObservation> AcceptedEnergyObservations,
+    IReadOnlyCollection<MarketActorRecipientV1> MarketActorRecipients)
+    : IInputParameterDto
+{
+    public record AcceptedEnergyObservation(
+        int Position,
+        decimal? EnergyQuantity,
+        Quality? QuantityQuality);
+}