+++ source/ProcessManager.Orchestrations.Tests/Fixtures/OrchestrationsAppFixture.cs
@@ -103,9 +103,6 @@ public class OrchestrationsAppFixture : IAsyncLifetime
     [NotNull]
     public IDurableClient? DurableClient { get; private set; }
 
-    [NotNull]
-    public string? ProcessManagerTopicName { get; private set; }
-
     public ServiceBusListenerMock EnqueueBrs021ForwardMeteredDataServiceBusListener { get; }
 
     public ServiceBusListenerMock EnqueueBrs023027ServiceBusListener { get; }
@@ -142,17 +139,8 @@ public class OrchestrationsAppFixture : IAsyncLifetime
 
         DurableClient = DurableTaskManager.CreateClient(TaskHubName);
 
-        // Create a shared Process Manager topic with subscriptions for both Orchestrations and Process Manager apps.
-        var processManagerTopicResourceBuilder = ServiceBusResourceProvider.BuildTopic("pm-topic");
-        OrchestrationsAppManager.ProcessManagerTopicResources.AddOrchestrationsAppSubscriptions(processManagerTopicResourceBuilder);
-        ProcessManagerAppManager.ProcessManagerTopicResources.AddSubscriptionsToTopicBuilder(processManagerTopicResourceBuilder);
-        var processManagerTopicResource = await processManagerTopicResourceBuilder.CreateAsync();
-
-        // Get resources from the created Process Manager topic for both Orchestrations and Process Manager apps.
-        var orchestrationsProcessManagerTopicResources = OrchestrationsAppManager.ProcessManagerTopicResources
-            .CreateFromTopic(processManagerTopicResource);
-        var processManagerAppProcessManagerTopicResources = ProcessManagerAppManager.ProcessManagerTopicResources
-            .CreateFromTopic(processManagerTopicResource);
+        // Process Manager Notify topic
+        await ProcessManagerAppManager.StartAsync();
 
         // Create EDI topic resources
         var ediTopicResources = await OrchestrationsAppManager.EdiTopicResources.CreateNew(ServiceBusResourceProvider);
@@ -177,10 +165,8 @@ public class OrchestrationsAppFixture : IAsyncLifetime
             integrationEventTopicResources.SharedTopic.Name,
             integrationEventTopicResources.Subscription.SubscriptionName);
 
-        await OrchestrationsAppManager.StartAsync(orchestrationsProcessManagerTopicResources, ediTopicResources, brs21TopicResources, integrationEventTopicResources);
-        await ProcessManagerAppManager.StartAsync(processManagerAppProcessManagerTopicResources);
-
-        ProcessManagerTopicName = orchestrationsProcessManagerTopicResources.ProcessManagerTopic.Name;
+        // Process Manager Orchestration Start topic
+        await OrchestrationsAppManager.StartAsync(ediTopicResources, brs21TopicResources, integrationEventTopicResources);
 
         EventHubListener = new EventHubListenerMock(
             new TestDiagnosticsLogger(),