+++ source/ProcessManager.Orchestrations.Tests/Integration/Processes/BRS_021/ForwardMeteredData/V1/MonitorOrchestrationUsingClientsScenario.cs
@@ -66,10 +66,7 @@ public class MonitorOrchestrationUsingClientsScenario : IAsyncLifetime
         var services = new ServiceCollection();
         services.AddInMemoryConfiguration(new Dictionary<string, string?>
         {
-            // Service bus client
-            [$"{ProcessManagerServiceBusClientOptions.SectionName}:{nameof(ProcessManagerServiceBusClientOptions.TopicName)}"]
-                = _fixture.ProcessManagerTopicName,
-            // Https client
+            // Process Manager HTTP client
             [$"{ProcessManagerHttpClientsOptions.SectionName}:{nameof(ProcessManagerHttpClientsOptions.ApplicationIdUri)}"]
                 = AuthenticationOptionsForTests.ApplicationIdUri,
             [$"{ProcessManagerHttpClientsOptions.SectionName}:{nameof(ProcessManagerHttpClientsOptions.GeneralApiBaseAddress)}"]
@@ -88,6 +85,12 @@ public class MonitorOrchestrationUsingClientsScenario : IAsyncLifetime
                 = _fixture.OrchestrationsAppManager.ProcessManagerEventhubName,
             [$"{ProcessManagerEventHubOptions.SectionName}:{nameof(ProcessManagerEventHubOptions.FullyQualifiedNamespace)}"]
                 = _fixture.IntegrationTestConfiguration.EventHubFullyQualifiedNamespace,
+
+            // Process Manager message client
+            [$"{ProcessManagerServiceBusClientOptions.SectionName}:{nameof(ProcessManagerServiceBusClientOptions.StartTopicName)}"]
+                = _fixture.OrchestrationsAppManager.ProcessManagerStartTopic.Name,
+            [$"{ProcessManagerServiceBusClientOptions.SectionName}:{nameof(ProcessManagerServiceBusClientOptions.NotifyTopicName)}"]
+                = _fixture.ProcessManagerAppManager.ProcessManagerNotifyTopic.Name,
         });
         services.AddAzureClients(
             builder => builder.AddServiceBusClientWithNamespace(_fixture.IntegrationTestConfiguration.ServiceBusFullyQualifiedNamespace));
@@ -143,7 +146,7 @@ public class MonitorOrchestrationUsingClientsScenario : IAsyncLifetime
         // Arrange
         var input = CreateMeteredDataForMeteringPointMessageInputV1();
 
-        var startCommand = new StartForwardMeteredDataCommandV1(
+        var startCommand = new ForwardMeteredDataCommandV1(
             new ActorIdentityDto(ActorNumber.Create(input.ActorNumber), ActorRole.FromName(input.ActorRole)),
             input,
             idempotencyKey: Guid.NewGuid().ToString());
@@ -258,10 +261,10 @@ public class MonitorOrchestrationUsingClientsScenario : IAsyncLifetime
         // TODO: Assert that the orchestration instance has been terminated successfully
     }
 
-    private static MeteredDataForMeteringPointMessageInputV1 CreateMeteredDataForMeteringPointMessageInputV1(
+    private static ForwardMeteredDataInputV1 CreateMeteredDataForMeteringPointMessageInputV1(
         bool withError = false)
     {
-        var input = new MeteredDataForMeteringPointMessageInputV1(
+        var input = new ForwardMeteredDataInputV1(
             "MessageId",
             Guid.NewGuid(),
             "1111111111111",
@@ -277,8 +280,7 @@ public class MonitorOrchestrationUsingClientsScenario : IAsyncLifetime
             "2024-12-02T23:00:00Z",
             "5790002606892",
             null,
-            new List<EnergyObservation>()
-            {
+            [
                 new("1", "112.000", "A04"),
                 new("2", "112.000", "A04"),
                 new("3", "112.000", "A04"),
@@ -302,7 +304,7 @@ public class MonitorOrchestrationUsingClientsScenario : IAsyncLifetime
                 new("22", "112.000", "A04"),
                 new("23", "112.000", "A04"),
                 new("24", "112.000", "A04"),
-            });
+            ]);
         return input;
     }