+++ source/ProcessManager.Orchestrations/Processes/BRS_021/ForwardMeteredData/V1/Activities/EnqueueRejectMessageActivity_Brs_021_V1.cs
@@ -60,6 +60,6 @@ internal class EnqueueRejectMessageActivity_Brs_021_V1(
 
     public record ActivityInput(
         OrchestrationInstanceId InstanceId,
-        MeteredDataForMeteringPointRejectedV1 RejectMessage,
+        ForwardMeteredDataRejectedV1 RejectMessage,
         Guid IdempotencyKey);
 }